+++ CHANGELOG.md
@@ -3,8 +3,14 @@
 * BREAKING CHANGE: The `m` parameter in in `Flay` and `Flayable` has been
   existentialized, to be chosen by the caller.
 
+* BREAKING CHANGE: We don't use `DefaultSignatures` for `flay = gflay` anymore.
+  This is very sud, but unfortunately type inferrence for the `c` type variable
+  in new `Flayable` instances doesn't work.
+
 * Added `Flayable1`, `trivial1`, `collect1`, `Terminal`, `GTerminal`, `zip1`,
   `zip`, `GFlay1`, `Record`, `GRecord`.
+
+* Removed `outer`.
 
 * Made compatible with GHC 8.2.1.